+++ CHANGELOG.md
@@ -6,12 +6,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+* `cleanCargoToml` can now be customized using the `pathsModifier` argument to
+  either add or remove additional paths to clean from the Cargo.toml file.
+
 ### Changed
 * **Breaking**: `mkCargoDerivation` now no longer automatically includes the
   `remapPathPrefixHook` as a nativeBuildInput. Unfortunately, this hook's
   implementation will lead to cache invalidation if built on a Nix
   implementation outside of NixOS (see the API documentation for more details).
   Thus using this hook is now *opt-in*.
+* **Breaking**: `cleanCargoToml` now removes unknown fields
+  (went from paths to remove to paths to keep)
 
 ## [0.22.0] - 2025-12-26